backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Document Processing
PyMuPDF==1.23.7
paddleocr==2.7.3
paddlepaddle==2.6.2
Pillow==10.1.0

# LLM & AI
openai>=1.30.0
langchain>=0.1.0
langchain-openai>=0.1.0
langsmith>=0.1.0
langgraph>=0.2.0
httpx>=0.27.0,<0.28.0

# Web Intelligence
google-search-results==2.4.2
beautifulsoup4==4.12.2
requests==2.31.0
ddgs>=9.11.0

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
shap==0.44.0
joblib==1.3.2

# Report Generation
python-docx==1.1.0

# CORS & Utils
aiofiles==23.2.1
